The cool JSON stuff I made was based on a misconception, so I got rid of that. Instead I made a funky reflection based pre/post hook thing that kinda achieves the same future proofing idea. Just don't forget to make fields that should be serialized transient. The model objects should extend Persistable.
The model (and the lock) now reside in a "warehouse", which is a factory that provides the same instance most of the time. I couldn't think of a better name. It also takes care of loading and saving (this allows the model to be entirely replaced). The warehouse is now the only singleton in the model package, and the access point for model stuff. Don't forget to monitor the lock when accessing the model.